Rewrote States and Game
CURRENTLY LAGGING - IN THE PROCESS OF BEING OPTIMIZED. Game is now the
only JFrame in the application and current state just draws to that
frame's canvas. This will make it really easy to build states for
Combat, Shops, etc later on. Also the components used in states are no
longer panels (no need).