Improved state manager
States can now be stacked on top of one another and pop of the stack
when no longer needed. For example, lets say we are on the map and we
encounter an enemy, we would add a combat state. Once we are done with
the combat state we would pop it off the stack and return the map state.