Updated Store State to be dynamic
Added a new StoreMenuUI that has a displayList and an itemList. Buying
an itemsremoves money from the counter and (should) add them to the
players inventory.
In the player constructor all items should be added (even if there are 0
being added) for the store to function correctly.
Set the name of StunBomb and RepairDrone in their constructers.