Added enemy ship generation
Enemy ships are now created when the grid is generated. They are also
spawned at a random enemy sector whenever the ship has moved a certain
distance. It still needs a little work. I also had to change The sector
state enum's name because it was the same as the sector state class, so
most of the sector classes were effected.